requirements.txt
dbt-snowflake
python-dotenv
pandas